+++ Resources/Constants.py
@@ -87,9 +87,9 @@ class Constants:
     # Payload Location
     # OpenCore
     @property
-    def opencore_zip_source(self): return self.payload_path / Path(f"OpenCore/OpenCore-{self.opencore_build}-v{self.opencore_version}.zip")
+    def opencore_zip_source(self): return self.payload_path / Path(f"OpenCore/OpenCore-{self.opencore_build}.zip")
     @property
-    def plist_template(self): return self.payload_path / Path(f"Config/v{self.opencore_version}/config.plist")
+    def plist_template(self): return self.payload_path / Path(f"Config/config.plist")
 
     # ACPI
     @property
@@ -155,9 +155,9 @@ class Constants:
     @property
     def build_path(self): return self.current_path / Path("Build-Folder/")
     @property
-    def opencore_release_folder(self): return self.build_path / Path(f"OpenCore-{self.opencore_build}-v{self.opencore_version}")
+    def opencore_release_folder(self): return self.build_path / Path(f"OpenCore-{self.opencore_build}")
     @property
-    def opencore_zip_copied(self): return self.build_path / Path(f"OpenCore-{self.opencore_build}-v{self.opencore_version}.zip")
+    def opencore_zip_copied(self): return self.build_path / Path(f"OpenCore-{self.opencore_build}.zip")
 
     @property
     def oc_folder(self): return self.opencore_release_folder / Path("EFI/OC/")