+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Increment AppleBackLightFixup v1.0.1
   - Add panel type F10T9cde
 - Enhance HDMI audio support on Mac Pros and Xserves
+- Strip unused kext entries during build
 
 ## 0.0.22
 - Add ExFat support for models missing driver